# perf(turbopack): improve app structure performance