### What?

Improves the performance of loader tree creation.

The main change is `directory_tree_to_loader_tree` now being completely synchronous (ignoring the setup).
This improves performance a lot (at least for the initial build), as it's a recursive function, which was cached at every level before.
It also makes querying the loader tree easier, as not everything needs to be an async turbo tasks function now.


#### `get_entrypoints` Performance
|                | Initial         | Update |
| --------- | ---------- | ------- |
| `canary` | `1300ms` | `20ms` |
| PR           | `260ms`  | `50ms` |